% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{bind_attr}
\alias{bind_attr}
\title{Bind Any Attribute}
\usage{
bind_attr(tag, name, expr)
}
\arguments{
\item{tag}{An htmltools tag.}

\item{name}{The attribute name (e.g. \code{"href"}, \code{"src"}).}

\item{expr}{A JavaScript expression string.}
}
\value{
The modified tag.
}
\description{
Dynamically binds any HTML attribute to a JavaScript expression.
}
\examples{
shiny::tags$a() |>
  bind_attr("href", "model.url") |>
  bind_text("model.linkText")

}
