% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{bind_checked}
\alias{bind_checked}
\title{Bind Checked State}
\usage{
bind_checked(tag, expr)
}
\arguments{
\item{tag}{An htmltools tag.}

\item{expr}{A JavaScript expression string referencing the model
(e.g. \code{"model.count"}).}
}
\value{
The modified tag.
}
\description{
Keeps a checkbox or radio input's checked state in sync with a
JavaScript expression.
}
\examples{
shiny::tags$input(type = "checkbox") |>
  bind_checked("model.agree")

}
