% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{bind_text}
\alias{bind_text}
\title{Bind Text Content}
\usage{
bind_text(tag, expr)
}
\arguments{
\item{tag}{An htmltools tag.}

\item{expr}{A JavaScript expression string referencing the model
(e.g. \code{"model.count"}).}
}
\value{
The modified tag.
}
\description{
Sets the element's text content from a JavaScript expression.
}
\examples{
shiny::tags$span() |> bind_text("model.count")

}
