% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{effect_send}
\alias{effect_send}
\title{Create a Custom Message Effect}
\usage{
effect_send(channel, data)
}
\arguments{
\item{channel}{The custom message channel name.}

\item{data}{The data to send (will be serialized to JSON).}
}
\value{
An \code{mvu_effect} object.
}
\description{
Describes a \link[shiny:session]{shiny::session}\verb{$sendCustomMessage()} call. Use this to
send data to client-side JavaScript handlers registered via
\code{extra_channels} in \code{\link[=mvu_page]{mvu_page()}} or \code{\link[=mvu_bridge_js]{mvu_bridge_js()}}.
}
\examples{
effect_send("clipboard", "text to copy")

}
