% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{effects}
\alias{effects}
\title{Combine Multiple Effects}
\usage{
effects(...)
}
\arguments{
\item{...}{Effect objects.}
}
\value{
An \code{mvu_effects} list.
}
\description{
Groups several effects into a single object that \code{\link[=mvu_result]{mvu_result()}} will
flatten. This is purely syntactic convenience.
}
\examples{
combined <- effects(
  effect_notify("Done"),
  effect_send("log", list(event = "done"))
)
mvu_result(list(status = "done"), combined)

}
