% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{is_mvu_result}
\alias{is_mvu_result}
\title{Test Whether an Object is an MVU Result}
\usage{
is_mvu_result(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
\code{TRUE} if \code{x} inherits from \code{"mvu_result"}.
}
\description{
Test Whether an Object is an MVU Result
}
