% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_set_in}
\alias{list_set_in}
\title{Set a Value at a Nested Path}
\usage{
list_set_in(.l, .path, .value)
}
\arguments{
\item{.l}{A list.}

\item{.path}{A character vector of keys forming the path.}

\item{.value}{The value to set.}
}
\value{
A new list with the value set at the given path.
}
\description{
Returns a modified copy of a list with \code{.value} placed at the location
specified by \code{.path}. Missing intermediate keys are created as empty
lists. Setting a value to \code{NULL} removes that key (standard R list
semantics).
}
\examples{
model <- list(a = list(b = 1))
list_set_in(model, c("a", "b"), 99)
list_set_in(model, c("a", "c", "d"), "new")
list_set_in(list(), c("x", "y"), 42)

}
