% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{mvu_button}
\alias{mvu_button}
\title{Action Button}
\usage{
mvu_button(label, msg, value = NULL, class = "btn btn-primary", ...)
}
\arguments{
\item{label}{The button label.}

\item{msg}{Message type to dispatch.}

\item{value}{Optional static value sent with the message.}

\item{class}{CSS classes. Defaults to Bootstrap primary button.}

\item{...}{Additional HTML attributes.}
}
\value{
A \verb{<button>} tag.
}
\description{
A \verb{<button>} that dispatches an MVU message on click.
}
\examples{
mvu_button("Click me", msg = "clicked")
mvu_button("+1", msg = "add", value = 1)

}
