% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{mvu_checkbox}
\alias{mvu_checkbox}
\title{Create a Message-Dispatching Checkbox}
\usage{
mvu_checkbox(label, msg, checked_expr, ...)
}
\arguments{
\item{label}{Label text displayed next to the checkbox.}

\item{msg}{Character string specifying the message type to dispatch.}

\item{checked_expr}{A JavaScript expression string for Alpine.js
\code{x-bind:checked}, typically \code{"model.isChecked"}.}

\item{...}{Additional HTML attributes passed to \link[shiny:reexports]{shiny::tags}\verb{$input}.}
}
\value{
A \verb{<div>} containing the checkbox input and label.
}
\description{
Generates a Bootstrap-styled checkbox that dispatches an MVU message on
change, sending the checked state as a boolean.
}
\examples{
mvu_checkbox(
  label = "Enable notifications",
  msg = "toggle_notifications",
  checked_expr = "model.notifications"
)

}
