% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{mvu_checkbox_input}
\alias{mvu_checkbox_input}
\title{Checkbox Input}
\usage{
mvu_checkbox_input(label, msg, bind = NULL, class = "form-check", ...)
}
\arguments{
\item{label}{Label text.}

\item{msg}{Message type to dispatch.}

\item{bind}{Alpine.js expression for the checked state.}

\item{class}{CSS classes for the wrapper div.}

\item{...}{Additional HTML attributes on the \verb{<input>}.}
}
\value{
A \verb{<div>} containing the checkbox and label.
}
\description{
A Bootstrap checkbox that dispatches its checked state as an MVU message.
}
\examples{
mvu_checkbox_input("I agree", msg = "toggle_agree", bind = "model.agree")

}
