% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{mvu_text}
\alias{mvu_text}
\title{Text Span}
\usage{
mvu_text(ref)
}
\arguments{
\item{ref}{A model reference or \code{\link[=raw_js]{raw_js()}} expression.}
}
\value{
An htmltools \verb{<span>} tag.
}
\description{
Creates a \verb{<span>} element whose text content is bound to a model
expression. For binding text to other element types, use \code{\link[=mvu_tag]{mvu_tag()}}
with the \code{text} parameter.
}
\examples{
m <- model_ref()
mvu_text(m$count)
mvu_text(raw_js("'Hello, ' + model.name"))

}
