% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{on_check}
\alias{on_check}
\title{Alpine.js Change Event Handler (Boolean Value)}
\usage{
on_check(msg)
}
\arguments{
\item{msg}{Character string specifying the message type.}
}
\value{
A named list with one \verb{@change} element.
}
\description{
Returns a named list with a \verb{@change} attribute that dispatches the
element's checked state as an MVU message. Use with checkboxes.
}
\examples{
shiny::tags$input(type = "checkbox", !!!on_check("toggle_dark_mode"))

}
