% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{on_click}
\alias{on_click}
\title{Alpine.js Click Event Handler}
\usage{
on_click(msg, value = NULL)
}
\arguments{
\item{msg}{Character string specifying the message type.}

\item{value}{An optional static R value sent with the message.}
}
\value{
A named list with one \verb{@click} element.
}
\description{
Returns a named list with an \verb{@click} attribute that dispatches an MVU
message. Splice into any tag with \verb{!!!}.
}
\examples{
shiny::tags$button(!!!on_click("increment"), "Click me")
shiny::tags$button(!!!on_click("set_page", value = 2), "Page 2")

}
