% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{on_click_expr}
\alias{on_click_expr}
\title{Alpine.js Click Event Handler with JavaScript Expression}
\usage{
on_click_expr(msg, value_expr)
}
\arguments{
\item{msg}{Character string specifying the message type.}

\item{value_expr}{A JavaScript expression string.}
}
\value{
A named list with one \verb{@click} element.
}
\description{
Like \code{\link[=on_click]{on_click()}} but accepts a JavaScript expression for the value,
evaluated at click time.
}
\examples{
shiny::tags$button(!!!on_click_expr("submit", "model.formData"), "Submit")

}
