% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{on_file}
\alias{on_file}
\title{File Input Event Handler}
\usage{
on_file(tag, msg)
}
\arguments{
\item{tag}{An htmltools tag.}

\item{msg}{Character string specifying the message type.}
}
\value{
The modified tag.
}
\description{
Dispatches file metadata (name, size, type) as a message when
files are selected.
}
\examples{
shiny::tags$input(type = "file") |> on_file("upload")

}
