% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{on_input_num}
\alias{on_input_num}
\title{Alpine.js Change Event Handler (Numeric Value)}
\usage{
on_input_num(msg, debounce = NULL)
}
\arguments{
\item{msg}{Character string specifying the message type.}

\item{debounce}{Optional debounce delay in milliseconds.}
}
\value{
A named list with one event attribute.
}
\description{
Returns a named list with a \verb{@change} (or debounced \verb{@input}) attribute
that dispatches the element's value coerced to a number.
}
\examples{
shiny::tags$input(type = "range", min = "0", max = "100",
  !!!on_input_num("set_volume"))

}
