% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{raw_js}
\alias{raw_js}
\title{Raw JavaScript Expression}
\usage{
raw_js(expr)
}
\arguments{
\item{expr}{A JavaScript expression string.}
}
\value{
An \code{mvu_raw_js} object.
}
\description{
Escape hatch for injecting a literal JavaScript expression into view
helpers. Use when the built-in helpers do not cover your case.
}
\examples{
raw_js("$event.target.value")
raw_js("model.items.length > 0")

}
