% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugger.R
\name{debugger_server}
\alias{debugger_server}
\title{Debugger Server Module}
\usage{
debugger_server(id, runtime, init)
}
\arguments{
\item{id}{Module ID matching the \code{\link[=debugger_ui]{debugger_ui()}} call.}

\item{runtime}{The runtime list produced when \code{debug = TRUE}.}

\item{init}{The init function used by the MVU component.}
}
\description{
Wires up the time-travel debugger to an MVU runtime. Used internally
when \code{debug = TRUE} is set on \code{\link[=mvu_module_server]{mvu_module_server()}}.
}
\keyword{internal}
