% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{mvu_button}
\alias{mvu_button}
\title{Create a Message-Dispatching Button}
\usage{
mvu_button(label, msg, value = NULL, class = "btn btn-primary", ...)
}
\arguments{
\item{label}{The button label text.}

\item{msg}{Character string specifying the message type to dispatch.}

\item{value}{An optional static value sent with the message. Character
values are quoted; other types are serialized to JSON.}

\item{class}{CSS class string for the button element.}

\item{...}{Additional HTML attributes passed to \link[shiny:reexports]{shiny::tags}\verb{$button}.}
}
\value{
A \verb{<button>} \link[shiny:reexports]{shiny::tags} element.
}
\description{
Generates a \verb{<button>} element wired to dispatch an MVU message via
Alpine.js \code{send()} on click.
}
\examples{
mvu_button("Click me", msg = "clicked")
mvu_button("+1", msg = "add", value = 1)
mvu_button("Save", msg = "save", value = "draft")

}
