% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinymvu-package.R
\docType{package}
\name{shinymvu-package}
\alias{shinymvu}
\alias{shinymvu-package}
\title{shinymvu: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
